requirements.txt
-e git+https://github.com/django/django.git@b4c80238a2b6c3f9fb7437d216082315c601055b#egg=django
WebOb==1.4
WebTest==2.0.15
beautifulsoup4==4.3.2
coverage==3.7.1
django-webtest==1.7.7
extras==0.0.3
fuzzywuzzy==0.2
gnureadline==6.3.3
ipdb==0.8
ipython==2.0.0
-e git+https://github.com/gabrielfalcao/lettuce.git@a63ed4b25f6928df959b6d33e5f534ad0036bbd6#egg=lettuce-master
lxml==3.3.5
nose==1.3.3
python-mimeparse==0.1.4
python-subunit==0.0.18
six==1.6.1
sure==1.2.7
testscenarios==0.4
testtools==0.9.35
waitress==0.8.9
wsgiref==0.1.2
psycopg2==2.5.3
gunicorn==18.0
dj-database-url==0.3.0
-e git+https://github.com/kennethreitz/dj-static.git@e79582483a976c76511244b3f236f7c68991e1ae#egg=dj-static
django-extensions==1.3.9
django-zurb-foundation==5.1.1
factory-boy==2.4.1
django-mailviews==0.6.3
djrill==1.1.1
